Refactor previous todos
